more sanity checks in case an out of sync config was used

+++ src/bat/fwunpack.py
@@ -1170,9 +1170,14 @@ def unpackCramfs(filename, offset, tempdir=None):
 def searchUnpackSquashfs(filename, tempdir=None, blacklist=[], offsets={}, envvars=None):
 	squashoffsets = []
 	for marker in fsmagic.squashtypes:
-		squashoffsets = squashoffsets + offsets[marker]
-	if squashoffsets == [] and offsets['squashfs7'] == []:
-		return ([], blacklist, [])
+		if offsets.has_key(marker):
+			squashoffsets = squashoffsets + offsets[marker]
+	if squashoffsets == []:
+		if offsets.has_key('squashfs7'):
+			if offsets['squashfs7'] == []:
+				return ([], blacklist, [])
+		else:
+			return ([], blacklist, [])
 
 	squashoffsets.sort()
 
@@ -1205,52 +1210,53 @@ def searchUnpackSquashfs(filename, tempdir=None, blacklist=[], offsets={}, envva
 	## it with unsquashfsRealtekLZMA
 	## TODO: don't suck in the file at once and see if it is possible to
 	## remove some duplicate code that is shared with the above code.
-	if offsets['squashfs7'] != []:
-		for offset in offsets['squashfs7']:
-			blacklistoffset = extractor.inblacklist(offset, blacklist)
-			if blacklistoffset != None:
-				continue
-			tmpdir = dirsetup(tempdir, filename, "squashfs", counter)
+	if offsets.has_key('squashfs7'):
+		if offsets['squashfs7'] != []:
+			for offset in offsets['squashfs7']:
+				blacklistoffset = extractor.inblacklist(offset, blacklist)
+				if blacklistoffset != None:
+					continue
+				tmpdir = dirsetup(tempdir, filename, "squashfs", counter)
 
-			sqshtmpdir = unpacksetup(tmpdir)
-			tmpfile = tempfile.mkstemp(dir=sqshtmpdir)
-			os.fdopen(tmpfile[0]).close()
+				sqshtmpdir = unpacksetup(tmpdir)
+				tmpfile = tempfile.mkstemp(dir=sqshtmpdir)
+				os.fdopen(tmpfile[0]).close()
 
-			sqshtmpfile = tempfile.mkstemp()
-			os.fdopen(sqshtmpfile[0]).close()
+				sqshtmpfile = tempfile.mkstemp()
+				os.fdopen(sqshtmpfile[0]).close()
 
-			## suck in the bytes up until the offset
-			sqshf = open(filename)
-			sqshf.seek(0)
-			sqshbytes = sqshf.read(offset)
+				## suck in the bytes up until the offset
+				sqshf = open(filename)
+				sqshf.seek(0)
+				sqshbytes = sqshf.read(offset)
 
-			## write out al the bytes until the offset
-			## then write 'sqsh'
-			sqshtmp = open(sqshtmpfile[1], 'w')
-			sqshtmp.write(sqshbytes + 'sqsh')
+				## write out al the bytes until the offset
+				## then write 'sqsh'
+				sqshtmp = open(sqshtmpfile[1], 'w')
+				sqshtmp.write(sqshbytes + 'sqsh')
 
-			## read the rest of the bytes from offset + 4
-			sqshf.seek(offset + 4)
-			sqshbytes = sqshf.read()
-			sqshf.close()
+				## read the rest of the bytes from offset + 4
+				sqshf.seek(offset + 4)
+				sqshbytes = sqshf.read()
+				sqshf.close()
 
-			## write them out
-			sqshtmp.write(sqshbytes)
-			sqshtmp.close()
+				## write them out
+				sqshtmp.write(sqshbytes)
+				sqshtmp.close()
 
-			## unpack, clean up, etc.
-			unpackFile(sqshtmpfile[1], offset, tmpfile[1], sqshtmpdir)
-			os.unlink(sqshtmpfile[1])
+				## unpack, clean up, etc.
+				unpackFile(sqshtmpfile[1], offset, tmpfile[1], sqshtmpdir)
+				os.unlink(sqshtmpfile[1])
 
-			retval = unpackSquashfsRealtekLZMA(tmpfile[1], offset, tmpdir)
-			os.unlink(tmpfile[1])
-			if retval != None:
-				(res, squashsize) = retval
-				diroffsets.append((res, offset, squashsize))
-				blacklist.append((offset,offset+squashsize))
-				counter = counter + 1
-			else:
-				os.rmdir(tmpdir)
+				retval = unpackSquashfsRealtekLZMA(tmpfile[1], offset, tmpdir)
+				os.unlink(tmpfile[1])
+				if retval != None:
+					(res, squashsize) = retval
+					diroffsets.append((res, offset, squashsize))
+					blacklist.append((offset,offset+squashsize))
+					counter = counter + 1
+				else:
+					os.rmdir(tmpdir)
 	return (diroffsets, blacklist, [])
 
 ## wrapper around all the different squashfs types
